Task2/requirements.txt
assemblyai
gradio
python-dotenv